Return (potentially) partial buffers on error.
In some cases, servers will response with an error, and then immediately
close the connection. In those cases, we can still recover some
information to be parsed for better error handling downstream.

One notable example is rate limiting. The current behaviour makes it
appear to be a connection issue (connection reset by peer). However, if
we inspect the buffer, we notice a rate limiting message. If we in turn
pass this to the whoisparser, we can treat it like other generic rate
limiting errors, and act accordingly.